requirements.txt
Flask==2.1.3
Pillow==9.4.0
PyMuPDF==1.21.1
chromadb==0.2.2
openai==0.27.0
sentence-transformers==2.2.2
torch==2.0.0
transformers==4.30.2
python-dotenv==1.0.0
